Phase 2 commit 3: extract http/{responses,body,router}.ts
- http/responses.ts: sendJson, sendError. Sole sanctioned response
  shape; handlers no longer call res.writeHead directly.
- http/body.ts: readBody (JSON), readTextBody, readBinaryBody.
- http/router.ts: Route, RouteHandler, defineRoute, dispatch. The
  dispatch loop in api.ts collapses to a one-line call.

api.ts loses the inline router/dispatch/body/response code that was
mixed in with handler logic; what remains is the route table and
the per-route handler closures.

No behaviour change. Server tests (33) green.
